performance(core): optimize browse auctions
