+++ mobile-app/app/screens/AppNavigator/screens/Dex/components/ButtonGroup.tsx
@@ -11,7 +11,6 @@ interface ButtonGroupProps {
 interface Buttons {
   id: string
   label: string
-  widthPercentage?: BigNumber
   handleOnPress: () => void
 }
 
@@ -30,7 +29,7 @@ export function ButtonGroup (props: ButtonGroupProps): JSX.Element {
             label={button.label}
             onPress={button.handleOnPress}
             isActive={props.activeButtonGroupItem === button.id}
-            width={button.widthPercentage ?? buttonWidth}
+            width={buttonWidth}
             key={button.id}
             testID={`${props.testID}_${button.id}`}
           />
